Conflicts:
	app/models/RequestData.scala

+++ app/models/Environment.scala
@@ -272,7 +272,7 @@ object Environment {
           gcal.setTimeInMillis(minDate.getTime + UtilDate.v1d)
           val maxDate = gcal.getTime
 
-          val result = RequestData.loadAvgResponseTimesByAction(e._1.toLong, minDate, maxDate, false)
+          val result = RequestData.loadAvgResponseTimesByAction(e._1, minDate, maxDate, false)
           result.foreach{(r) =>
             Logger.debug("env:" + e._2 + " SoapAction:"+r._1+ " timeAverage:" + r._2)
             RequestData.insertStats(e._1.toLong, r._1, minDate, r._2)
